diffs: UnknownAsNull function
This helper takes a value that might contain unknowns and returns a new
value with them all replaced with nulls. This is useful for eliminating
any stray unknowns in a context where everything ought to be known by now,
e.g. as input into cty JSON serialization where unknown values cannot be
represented.